+++ src/sdk/CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Fixed `FieldLink` handling when `ContentTypeHub` was used [jansenbe - Bert Jansen]
 - Fixed `AddTaxonomy` field methods when `ContentTypeHub` was used [jansenbe - Bert Jansen]
 - Handle the CSOM limitation of not being able to set an empty `RestrictedAccessControlGroups` property on `ISiteCollectionProperties.Update` [jansenbe - Bert Jansen]
+- Setting the localized name of a TermSet was not possible #1188 [jansenbe - Bert Jansen]
 
 ## [1.9.0]